Integrate kube mgmt (#6)
* Add GOPROXY to Dockerfile

* Add build target to Makefile

* Add kube-mgmt to the deployment

* Update go modules

* Update Makefile to support redeploy on K3s

* Add HTTP handler for the redis enforcement

* Integrate with OPA

* Refactor

* Prevent empty tenant IDs from old tokens

* Add karavictl and self-sign functionality

* Focus RBAC on token group rather than tenant

+++ deploy/airgap-prepare.sh
@@ -14,13 +14,22 @@ INSTALL_SCRIPT=install.sh
 
 
 # Download install script
-curl -kL -o $K3S_INSTALL_SCRIPT https://get.k3s.io/
+if [[ ! -f $K3S_INSTALL_SCRIPT ]]
+then
+	curl -kL -o $K3S_INSTALL_SCRIPT https://get.k3s.io/
+fi
 
 # Download k3s
-curl -kL -o $K3S_BINARY  https://github.com/rancher/k3s/releases/download/v1.18.10%2Bk3s1/k3s
+if [[ ! -f $K3S_BINARY ]]
+then
+	curl -kL -o $K3S_BINARY  https://github.com/rancher/k3s/releases/download/v1.18.10%2Bk3s1/k3s
+fi
 
-# Download k3s images
-curl -kL -o $K3S_IMAGES_TAR https://github.com/rancher/k3s/releases/download/v1.18.10%2Bk3s1/k3s-airgap-images-$ARCH.tar
+if [[ ! -f $K3S_IMAGES_TAR ]]
+then
+	# Download k3s images
+	curl -kL -o $K3S_IMAGES_TAR https://github.com/rancher/k3s/releases/download/v1.18.10%2Bk3s1/k3s-airgap-images-$ARCH.tar
+fi
 
 # Save all referenced images into a tarball
 grep "image: " deployment.yaml | awk -F' ' '{ print $2 }' | xargs docker save -o $CRED_SHIELD_IMAGES_TAR
